Fully-functional pause menu
Pause menu is completed and mechanics are working as expected. I will
update the buttons so that they match the Title Screen buttons, but
other than that everything is working. For future reference, in order
for anything in a Canvas to be interactable, the Canvas must have the
GraphicRaycaster component attached to it.