basic Enemy prefab change
I added the Regular Wolf sprites to the Enemy prefab and changed the
animation in order to minimize some of the strange results I've been
getting. The problem with it right now is that when they walk their body
parts are detached from each other. That's because I made them in a 2D
view in order to line them up correctly. Also, when the Enemy-type now
walks into the FinishRegion, it doesn't get destroyed. But when it gets
shot by a bullet it does.