Enemy prefab and scene changes
Made changes to the Enemy prefab in regards to simplifying the walking
animation so the neck and jaw don't look detached when they were
rotating (only the legs are moving now) and changed the entire scene to
orthographic view so our animations for our enemies don't look messed
up. We will need to tweak player movement though so that he moves in a
more diagonal line in order to match the angle of the fence and house.
Also commented out the Debug.Log statement in the EnemyWave script.